Add work info screen and delete functionality
Introduced a new screen `MyWorksInfoScreen` to display detailed information about a user's work. This includes a video player using ExoPlayer to play the generated video. Added functionality to delete a work from the local database.

Updated the `MyWorksScreen` to navigate to the new info screen when a work is clicked. Modified `MyWorksViewModel` and related classes to handle navigation and deletion.

Also included a new delete icon and updated the database version.